feat: move BuiltinVariableNameStrategy into core, in order to share at CLI and API
