+++ jsgenerator-cli/src/main/java/com/osscameroon/jsgenerator/cli/internal/CommandDefault.java
@@ -1,6 +1,6 @@
 package com.osscameroon.jsgenerator.cli.internal;
 
-import com.osscameroon.jsgenerator.cli.BuiltinVariableNameStrategy;
+import com.osscameroon.jsgenerator.core.BuiltinVariableNameStrategy;
 import com.osscameroon.jsgenerator.cli.Command;
 import com.osscameroon.jsgenerator.cli.Valid;
 import com.osscameroon.jsgenerator.core.Configuration;
@@ -20,7 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import static com.osscameroon.jsgenerator.cli.BuiltinVariableNameStrategy.TYPE_BASED;
+import static com.osscameroon.jsgenerator.core.BuiltinVariableNameStrategy.TYPE_BASED;
 import static com.osscameroon.jsgenerator.core.OutputStreamResolver.*;
 import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
